Implement update_apps
Of note, when Git borks and `box update-apps` prompts for a response
from the user, there's no way that I've found to capture that input and
send it back to `box update-apps`. :shrug:. For now, running the whole
command non-interactively let's things fail with actionable error
messages so that the user can carry on.