Start moving this thing towards a proper gem
Take a lot of inspiration from TTY's `teletype` scaffolding.
Unfortunately, it's a pain to actually use `teletype` right now because
of issue's it has with bundler v2. 😞 Setup a dir structure for
organizing a gem, with an executable in `exe`, a `commands` folder for
commands that can inherit from a base `Command` class that handles
lazily `require` -ing and instantiating TTY components.