Implement posting to Slack
Refine and simplify the approach a bit. With Slack as a command, we can
allow options for passing a channel and message. We can invoke sending a
message from Start and End with ease, passing along a message as needed.
We can refine futher in the future if the need arises. I can imagine
allowing a prompt for channel, skipping signing off or prompting for a
signoff message when ending the day.